Fix comments (move extension to separate file, group all private methods under a private extension, private IBOutlets)
